requirements.txt
prokka
cd-hit
cd-hit-2D
Biopython
